Change the implementation of climate change impacts on water yield via year-to-year checking

+++ luto/tools/write.py
@@ -99,16 +99,24 @@ def write_data(data: Data):
     # Write outputs for each year
     jobs = [delayed(write_output_single_year)(data, yr, path_yr, None) for (yr, path_yr) in zip(years, paths)]
 
+    # Check if the simulation is complete by comparing the last year in the simulation with the target year
+    complete_simulation = max([int(i[-4:]) for i in os.listdir(data.path) if 'out' in i]) == max(years)
+    
     # Write the area/quantity comparison between base-year and target-year for the timeseries mode
-    begin_end_path = f"{data.path}/begin_end_compare_{years[0]}_{years[-1]}"
-    jobs += [delayed(write_output_single_year)(data, years[-1], f"{begin_end_path}/out_{years[-1]}", years[0])] if settings.MODE == 'timeseries' else []
+    if complete_simulation:
+        begin_end_path = f"{data.path}/begin_end_compare_{years[0]}_{years[-1]}"
+        jobs += [delayed(write_output_single_year)(data, years[-1], f"{begin_end_path}/out_{years[-1]}", years[0])] if settings.MODE == 'timeseries' else []
+    else:
+        print(f'''The target year is not the last year in the simulation!
+                  Only Writing the avaliable outputs ({years[0]}-{years[-1]}) to output directory.\n''')
 
     # Parallel write the outputs for each year
     num_jobs = min(len(jobs), settings.WRITE_THREADS) if settings.PARALLEL_WRITE else 1   # Use the minimum between jobs_num and threads for parallel writing
     Parallel(n_jobs=num_jobs)(jobs)
 
     # Copy the base-year outputs to the path_begin_end_compare
-    shutil.copytree(f"{data.path}/out_{years[0]}", f"{begin_end_path}/out_{years[0]}", dirs_exist_ok = True) if settings.MODE == 'timeseries' else None
+    if complete_simulation:
+        shutil.copytree(f"{data.path}/out_{years[0]}", f"{begin_end_path}/out_{years[0]}", dirs_exist_ok = True) if settings.MODE == 'timeseries' else None
     
     # Create the report HTML and png maps
     TIF2MAP(data.path) if settings.WRITE_OUTPUT_GEOTIFFS else None
